requirements.txt
# Core dependencies for the Agentic RAG system
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.25.2
xarray==2023.11.0
tqdm==4.66.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning and AI
openai==1.3.8
scikit-learn==1.3.2
sentence-transformers==2.2.2

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data processing
netcdf4==1.6.5
pyarrow==14.0.1

# Caching and performance
redis==5.0.1
celery==5.3.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
mypy==1.7.1
